define interfaces for LinkedList and Set
LinkedList needed to be a pointer type, while Set could be a reference
type. Rather than changing the implementation of one or the other to be
more internally consistent, we define interfaces for each collection
type and the constructor returns the interface type. This also allows us
to privatize the implementation types, such that we don't need to worry
about them being constructed in a nonstandard way. (Constructing a raw
Set{} is gonna lead to a bad time.)